+++ app/models/display_ad.rb
@@ -1,5 +1,5 @@
 class DisplayAd < ApplicationRecord
-  include TagListValidateable
+  include Taggable
   acts_as_taggable_on :tags
   resourcify
 
@@ -32,8 +32,19 @@ class DisplayAd < ApplicationRecord
                              search: "%#{term}%"
                      }
 
-  def self.for_display(area, user_signed_in)
+  def self.for_display(area, user_signed_in, article_tags = [])
     relation = approved_and_published.where(placement_area: area).order(success_rate: :desc)
+
+    if article_tags.any?
+      display_ads_with_no_tags = relation.where(cached_tag_list: "")
+      display_ads_with_targeted_article_tags = relation.cached_tagged_with_any(article_tags)
+
+      relation = display_ads_with_no_tags.or(display_ads_with_targeted_article_tags)
+    end
+
+    if article_tags.blank?
+      relation = relation.where(cached_tag_list: "")
+    end
 
     relation = if user_signed_in
                  relation.where(display_to: %w[all logged_in])
