+++ app/views/articles/show.html.erb
@@ -191,7 +191,7 @@
       </article>
 
       <% cache("article-bottom-content-#{@article.id}-#{user_signed_in?}-#{(@organization || @user).latest_article_updated_at}", expires_in: 18.hours) do %>
-        <% @display_ad = DisplayAd.for_display("post_comments", user_signed_in?) %>
+        <% @display_ad = DisplayAd.for_display("post_comments", user_signed_in?, @article.decorate.cached_tag_list_array) %>
         <% if @display_ad %>
           <div class="crayons-card crayons-card--secondary p-4 crayons-sponsorship-article text-styles"
                data-display-unit data-id="<%= @display_ad.id %>"
