+++ db/migrate/20221031133257_add_cached_tag_list_to_display_ads.rb
@@ -1,0 +1,5 @@
+class AddCachedTagListToDisplayAds < ActiveRecord::Migration[7.0]
+  def change
+    add_column :display_ads, :cached_tag_list, :string
+  end
+end
