+++ db/migrate/20221031133328_add_index_to_display_ads_cached_tag_list.rb
@@ -1,0 +1,9 @@
+class AddIndexToDisplayAdsCachedTagList < ActiveRecord::Migration[7.0]
+  disable_ddl_transaction!
+
+  def change
+    # From <https://www.postgresql.org/docs/11/pgtrgm.html#id-1.11.7.40.7>
+    # We need a GIN index on `cached_tag_list` to speed up `LIKE` queries
+    add_index :display_ads, :cached_tag_list, using: :gin, opclass: :gin_trgm_ops, algorithm: :concurrently
+  end
+end
