+++ spec/models/article_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe Article, type: :model do
 
   include_examples "#sync_reactions_count", :article
   it_behaves_like "UserSubscriptionSourceable"
+  it_behaves_like "Taggable"
 
   describe "validations" do
     it { is_expected.to belong_to(:collection).optional }
@@ -984,210 +985,6 @@ RSpec.describe Article, type: :model do
     it "returns nothing if no tagged articles" do
       articles = described_class.search_optimized("godsdsdsdsgoo")
       expect(articles).to be_empty
-    end
-  end
-
-  describe ".cached_tagged_with" do
-    it "can search for a single tag" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      articles = described_class.cached_tagged_with("includeme")
-
-      expect(articles).to include included
-      expect(articles).not_to include excluded
-      expect(articles.to_a).to eq described_class.tagged_with("includeme").to_a
-    end
-
-    it "can search for a single tag when given a symbol" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      articles = described_class.cached_tagged_with(:includeme)
-
-      expect(articles).to include(included)
-      expect(articles).not_to include(excluded)
-      expect(articles.to_a).to eq(described_class.tagged_with("includeme").to_a)
-    end
-
-    it "can search for a single tag when given a Tag object" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      tag = Tag.find_by(name: :includeme)
-
-      articles = described_class.cached_tagged_with(tag)
-
-      expect(articles).to include included
-      expect(articles).not_to include excluded
-      expect(articles.to_a).to eq described_class.tagged_with("includeme").to_a
-    end
-
-    it "can search among multiple tags" do
-      included = [
-        create(:article, tags: "omg, wtf"),
-        create(:article, tags: "omg, lol"),
-      ]
-      excluded = create(:article, tags: "nope, excluded")
-
-      articles = described_class.cached_tagged_with("omg")
-
-      expect(articles).to include(*included)
-      expect(articles).not_to include excluded
-      expect(articles.to_a).to include(*described_class.tagged_with("omg").to_a)
-    end
-
-    it "can search for multiple tags" do
-      included = create(:article, tags: "includeme, please, lol")
-      excluded_partial_match = create(:article, tags: "excluded, please")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      articles = described_class.cached_tagged_with(%w[includeme please])
-
-      expect(articles).to include included
-      expect(articles).not_to include excluded_partial_match
-      expect(articles).not_to include excluded_no_match
-      expect(articles.to_a).to eq described_class.tagged_with(%w[includeme please]).to_a
-    end
-
-    it "can search for multiple tags passed as an array of symbols" do
-      included = create(:article, tags: "includeme, please, lol")
-      excluded_partial_match = create(:article, tags: "excluded, please")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      articles = described_class.cached_tagged_with(%i[includeme please])
-
-      expect(articles).to include(included)
-      expect(articles).not_to include(excluded_partial_match)
-      expect(articles).not_to include(excluded_no_match)
-      expect(articles.to_a).to eq(described_class.tagged_with(%i[includeme please]).to_a)
-    end
-
-    it "can search for multiple tags passed as an array of Tag objects" do
-      included = create(:article, tags: "includeme, please, lol")
-      excluded_partial_match = create(:article, tags: "excluded, please")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      tags = Tag.where(name: %i[includeme please]).to_a
-      articles = described_class.cached_tagged_with(tags)
-
-      expect(articles).to include(included)
-      expect(articles).not_to include(excluded_partial_match)
-      expect(articles).not_to include(excluded_no_match)
-      expect(articles.to_a).to eq(described_class.tagged_with(%i[includeme please]).to_a)
-    end
-  end
-
-  describe ".cached_tagged_with_any" do
-    it "can search for a single tag" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      articles = described_class.cached_tagged_with_any("includeme")
-
-      expect(articles).to include included
-      expect(articles).not_to include excluded
-      expect(articles.to_a).to eq described_class.tagged_with("includeme", any: true).to_a
-    end
-
-    it "can search for a single tag when given a symbol" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      articles = described_class.cached_tagged_with_any(:includeme)
-
-      expect(articles).to include(included)
-      expect(articles).not_to include(excluded)
-      expect(articles.to_a).to eq(described_class.tagged_with("includeme", any: true).to_a)
-    end
-
-    it "can search for a single tag when given a Tag object" do
-      included = create(:article, tags: "includeme")
-      excluded = create(:article, tags: "lol, nope")
-
-      tag = Tag.find_by(name: :includeme)
-      articles = described_class.cached_tagged_with_any(tag)
-
-      expect(articles).to include(included)
-      expect(articles).not_to include(excluded)
-      expect(articles.to_a).to eq(described_class.tagged_with("includeme", any: true).to_a)
-    end
-
-    it "can search among multiple tags" do
-      included = [
-        create(:article, tags: "omg, wtf"),
-        create(:article, tags: "omg, lol"),
-      ]
-      excluded = create(:article, tags: "nope, excluded")
-
-      articles = described_class.cached_tagged_with_any("omg")
-      expected = described_class.tagged_with("omg", any: true).to_a
-
-      expect(articles).to include(*included)
-      expect(articles).not_to include excluded
-      expect(articles.to_a).to include(*expected)
-    end
-
-    it "can search for multiple tags" do
-      included = create(:article, tags: "includeme, please, lol")
-      included_partial_match = create(:article, tags: "includeme, omg")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      articles = described_class.cached_tagged_with_any(%w[includeme please])
-      expected = described_class.tagged_with(%w[includeme please], any: true).to_a
-
-      expect(articles).to include included
-      expect(articles).to include included_partial_match
-      expect(articles).not_to include excluded_no_match
-
-      expect(articles.to_a).to include(*expected)
-    end
-
-    it "can search for multiple tags when given an array of symbols" do
-      included = create(:article, tags: "includeme, please, lol")
-      included_partial_match = create(:article, tags: "includeme, omg")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      articles = described_class.cached_tagged_with_any(%i[includeme please])
-      expected = described_class.tagged_with(%i[includeme please], any: true).to_a
-
-      expect(articles).to include(included)
-      expect(articles).to include(included_partial_match)
-      expect(articles).not_to include(excluded_no_match)
-
-      expect(articles.to_a).to include(*expected)
-    end
-
-    it "can search for multiple tags when given an array of Tag objects" do
-      included = create(:article, tags: "includeme, please, lol")
-      included_partial_match = create(:article, tags: "includeme, omg")
-      excluded_no_match = create(:article, tags: "excluded, omg")
-
-      tags = Tag.where(name: %i[includeme please]).to_a
-      articles = described_class.cached_tagged_with_any(tags)
-      expected = described_class.tagged_with(%i[includeme please], any: true).to_a
-
-      expect(articles).to include(included)
-      expect(articles).to include(included_partial_match)
-      expect(articles).not_to include(excluded_no_match)
-
-      expect(articles.to_a).to include(*expected)
-    end
-  end
-
-  describe ".not_cached_tagged_with_any" do
-    it "can exclude multiple tags when given an array of strings" do
-      included = create(:article, tags: "includeme")
-      excluded1 = create(:article, tags: "includeme, lol")
-      excluded2 = create(:article, tags: "includeme, omg")
-
-      articles = described_class
-        .cached_tagged_with_any("includeme")
-        .not_cached_tagged_with_any(%w[lol omg])
-
-      expect(articles).to include included
-      expect(articles).not_to include excluded1
-      expect(articles).not_to include excluded2
     end
   end
